Bug fix: Expected a JsonObject but was JsonPrimitive
Error Message : An exception occurred while executing the Java class. Expected a com.google.gson.JsonObject but was com.google.gson.JsonPrimitive; at path $.id -> [Help 1]

If create() in FlatteningTypeAdapterFactory.java is called with a primitive type, we will not create a TypeAdapter, instead we will return null meaning we will send the problem to a deafult Gson create method.